% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pwr_conj_test.R
\name{calc_K_conj_test}
\alias{calc_K_conj_test}
\title{Calculate required number of clusters per treatment group for a cluster randomized trial with co-primary endpoints using the conjunctive intersection-union test approach.}
\usage{
calc_K_conj_test(
  power,
  m,
  alpha = 0.05,
  beta1,
  beta2,
  varY1,
  varY2,
  rho01,
  rho02,
  rho1,
  rho2,
  r = 1,
  cv = 0,
  deltas = c(0, 0)
)
}
\arguments{
\item{power}{Desired statistical power in decimal form; numeric.}

\item{m}{Individuals per cluster; numeric.}

\item{alpha}{Type I error rate; numeric.}

\item{beta1}{Effect size for the first outcome; numeric.}

\item{beta2}{Effect size for the second outcome; numeric.}

\item{varY1}{Total variance for the first outcome; numeric.}

\item{varY2}{Total variance for the second outcome; numeric.}

\item{rho01}{Correlation of the first outcome for two different individuals in the same cluster; numeric.}

\item{rho02}{Correlation of the second outcome for two different individuals in the same cluster; numeric.}

\item{rho1}{Correlation between the first and second outcomes for two individuals in the same cluster; numeric.}

\item{rho2}{Correlation between the first and second outcomes for the same individual; numeric.}

\item{r}{Treatment allocation ratio - K2 = rK1 where K1 is number of clusters in experimental group; numeric.}

\item{cv}{Cluster variation parameter, set to 0 if assuming all cluster sizes are equal; numeric.}

\item{deltas}{Vector of non-inferiority margins, set to delta_1 = delta_2 = 0; numeric vector.}
}
\value{
A data frame of numerical values.
}
\description{
Allows user to calculate the required number of clusters per treatment group of a hybrid type 2 cluster randomized trial given a set of study design input values, including the statistical power, and cluster size. Uses the conjunctive intersection-union test approach.Code is adapted from "calSampleSize_ttestIU()" from https://github.com/siyunyang/coprimary_CRT written by Siyun Yang.
}
\examples{
calc_K_conj_test(power = 0.8, m = 300, alpha = 0.05, beta1 = 0.1, beta2 = 0.1, varY1 = 0.23, varY2 = 0.25, rho01 = 0.025, rho02 = 0.025, rho1 = 0.01, rho2  = 0.05)
}
