% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ncp_chi2.R
\name{calc_ncp_chi2}
\alias{calc_ncp_chi2}
\title{Find the non-centrality parameter corresponding to Type I error rate and statistical power}
\usage{
calc_ncp_chi2(alpha, power, df = 1)
}
\arguments{
\item{alpha}{Type I error rate; numeric.}

\item{power}{Desired statistical power in decimal form; numeric.}

\item{df}{Degrees of freedom; numeric.}
}
\value{
A number.
}
\description{
Allows user to find the corresponding non-centrality parameter for power analysis based on the Type I error rate, statistical power, and degrees of freedom.
}
\examples{
calc_ncp_chi2(alpha = 0.05, power = 0.8, df = 1)
}
